move PDF component to components
