+++ src/components/PDF/ItemsTable/TableBlankSpace.jsx
